refactor/server-actions: update markdown import paths
