+++ src/utils/markdown.ts
@@ -1,3 +1,4 @@
+"use server"
 import fs from "fs"
 import path from "path"
 import matter from "gray-matter"
